feat: move type CodeSnippet
