+++ apollo/apollo-client.ts
@@ -1,7 +1,7 @@
 import { ApolloClient, InMemoryCache } from '@apollo/client';
 
 const apolloClient = new ApolloClient({
-  uri: 'https://martacodes-it-strapi.up.railway.app/graphql',
+  uri: process.env.GRAPHQL_URL,
   cache: new InMemoryCache(),
   defaultOptions: {
     watchQuery: {
