+++ src/app/(public)/page.tsx
@@ -1,28 +1,33 @@
 import * as React from 'react';
 
+import { flattenToArray } from '@/lib/graphqlUtils';
+
 import Intro from '@/components/organisms/home/Intro';
 import Summary from '@/components/organisms/home/Summary';
 import Seo from '@/components/Seo';
 
-import { codeSnippetsQuery } from '@/queries/codeSnippets';
-import { CodeSnippet } from '@/sanityTypes/CodeSnippet';
+import { codeSnippetsQueryQL } from '@/queries/codeSnippets';
 
-import { sanityClient } from '../../../sanity/lib/client';
+import apolloClient from '../../../apollo/apollo-client';
 
-const getData = async () => {
-  // const homeData: HomeContent[] = await sanityClient.fetch(homeContentQuery);
-  const codeSnippets: CodeSnippet[] = await sanityClient.fetch(
-    codeSnippetsQuery
-  );
+import { CodeSnippet } from '@/types/CodeSnippet';
+
+async function queryCodeSnippets() {
+  const { data } = await apolloClient.query({ query: codeSnippetsQueryQL });
+
+  return flattenToArray<CodeSnippet>(data.codeSnippets);
+}
+
+const queryData = async () => {
+  const codeSnippets: CodeSnippet[] = await queryCodeSnippets();
 
   return {
     codeSnippets,
-    // homeContent: homeData[0],
   };
 };
 
 const HomePage = async () => {
-  const { codeSnippets } = await getData();
+  const { codeSnippets } = await queryData();
 
   return (
     <main className='min-h-main'>
