+++ src/types/CodeSnippet.ts
@@ -1,0 +1,5 @@
+export interface CodeSnippet {
+  id: string;
+  code: string;
+  language: string;
+}
